do-cmake_minis: Remove explicit en/disable pnetcdf
For some reason, explicitly enabling PNetCDF breaks the ability to find
it the normal way.  So, rely on the TriBITS logic that I presume checks
to see if NetCDF is on, and MPI is on to automatically enable it if
needed (this is what we do in the normal builds, so if it breaks, it
will break there too).  In the MPI=OFF case, it warns you that the
PNetCDF_ROOT is unused, but again, if it were to try and use PNetCDF, it
should fail at compile time when it tries to include MPI headers.

User Support Ticket(s) or Story Referenced: SPAR-747