Enable/disable only ShyLU_NodeHTS and never ShyLU_DD (CDOFA-110)
Just enable or disable the TriBITS subpackage ShyLU_NodeHTS as that is all of
the ShyLU software that SPARC uses.

The subpackage ShyLU_NodeTacho must be explicitly disabled because Amesos2 has
an optional dependnecy on ShyLU_NodeTacho and therefore we must explicitly
disable it if we don't want it built.

SPARC never uses any code in the TriBITS ShyLU_DD package.  It was only
getting enabled to avoid a configure error when trying to enable the psuedo
package ShyLU that has a required depenency on the TriBITS package ShyLU_DD.
So just don't enable the psuedo package ShyLU.  Problem solved.

This has reduced these do-cmake* scripts by 5 lines and more directly says
that SPARC needs (and does not want).  And we avoid enabling, building, and
installing software that is never used by SPARC.  Less configure arguments,
less software being built, and more direct statement of what SPARC is actually
using from Trilinos => Win, win win.

NOTE: Once we can switch SPARC over to using the ATDM Trilinos configuration,
we will not have to make the same modifications in 20 files for a simple
changes like this.  Basically, ShyLU_NodeHTS is enabled in every build except
for CUDA builds.  We can express that logic must more compactly with *.cmake
options files.